+++ TimeDet/TimeDetHit.cxx
@@ -19,8 +19,9 @@
 using std::cout;
 using std::endl;
 
-Double_t speedOfLight =
-    TMath::C() * 100. / 1000000000.0;  // from m/sec to cm/ns
+namespace {
+constexpr Double_t speedOfLight = 29.9792458;  // TMath::C() * 100 / 1e9, cm/ns
+}  // namespace
 
 // -----   Default constructor   --------------
 TimeDetHit::TimeDetHit() : ShipHit() { flag = true; }